rtm: big refactor
