+++ packages/rtm-server/test/server.test.ts
@@ -40,70 +40,6 @@ test('connects to lightning', async (t) => {
   // TODO connections to hte same socket.channel should share listners, so I think I can test the channel
 });
 
-test('workloop can be cancelled', async (t) => {
-  let count = 0;
-  let cancel;
-  const channel = mockChannel({
-    [CLAIM]: () => {
-      count++;
-      cancel();
-    },
-  });
-
-  cancel = startWorkloop(channel, () => {}, 1);
-
-  await sleep(100);
-  // A quirk of how cancel works is that the loop will be called a few times
-  t.assert(count <= 5);
-});
-
-test('workloop sends the attempts:claim event', (t) => {
-  return new Promise((done) => {
-    let cancel;
-    const channel = mockChannel({
-      [CLAIM]: () => {
-        t.pass();
-        done();
-      },
-    });
-    cancel = startWorkloop(channel, () => {});
-  });
-});
-
-test('workloop sends the attempts:claim event several times ', (t) => {
-  return new Promise((done) => {
-    let cancel;
-    let count = 0;
-    const channel = mockChannel({
-      [CLAIM]: () => {
-        count++;
-        if (count === 5) {
-          t.pass();
-          done();
-        }
-      },
-    });
-    cancel = startWorkloop(channel, () => {});
-  });
-});
-
-test('workloop calls execute if attempts:claim returns attempts', (t) => {
-  return new Promise((done) => {
-    let cancel;
-    const channel = mockChannel({
-      [CLAIM]: () => {
-        return [{ id: 'a' }];
-      },
-    });
-
-    cancel = startWorkloop(channel, (attempt) => {
-      t.deepEqual(attempt, { id: 'a' });
-      t.pass();
-      done();
-    });
-  });
-});
-
 // test('connects to websocket', (t) => {
 //   let didSayHello;
